+++ packages/aws-appsync-subscription-link/__tests__/link/realtime-subscription-handshake-link-test.ts
@@ -1,4 +1,4 @@
-import { AUTH_TYPE } from "aws-appsync-auth-link";
+import { AUTH_TYPE } from "@boxcc/aws-appsync-auth-link";
 import { execute } from "apollo-link";
 import gql from 'graphql-tag';
 import { AppSyncRealTimeSubscriptionHandshakeLink } from '../../src/realtime-subscription-handshake-link';