+++ packages/aws-appsync/__tests__/client.ts
@@ -34,7 +34,7 @@ beforeEach(() => {
         const { AWSAppSyncClient } = require('../src/client');
         ({ isOptimistic } = require("../src/link/offline-link"));
         ({ createHttpLink } = require("apollo-link-http"));
-        ({ Signer } = require("aws-appsync-auth-link"));
+        ({ Signer } = require("@boxcc/aws-appsync-auth-link"));
 
         factory = (opts) => {
             return new AWSAppSyncClient(opts);
@@ -710,7 +710,9 @@ describe("Offline enabled", () => {
                 name: 'Parent'
             },
             optimisticResponse: optimisticResponseParent,
-            update: (proxy, { data }) => {
+            update: <T = {
+                [key: string]: any;
+            }>(proxy, { data }) => {
                 proxy.writeQuery({
                     query: gql`query Bla($id: ID) {
                         getParent(id: $id) {
@@ -745,7 +747,9 @@ describe("Offline enabled", () => {
                 name: 'Child'
             },
             optimisticResponse: optimisticResponseChild,
-            update: (proxy, { data }) => {
+            update: <T = {
+                [key: string]: any;
+            }>(proxy, { data }) => {
                 proxy.writeQuery({
                     query: gql`query Ble($id: ID!) {
                         getChild(id: $id) {