+++ packages/aws-appsync-auth-link/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.2"></a>
+## [2.0.2](https://github.com/awslabs/aws-mobile-appsync-sdk-js/compare/aws-appsync-auth-link@2.0.1...aws-appsync-auth-link@2.0.2) (2020-04-15)
+
+
+
+
+**Note:** Version bump only for package aws-appsync-auth-link
+
 <a name="2.0.1"></a>
 ## [2.0.1](https://github.com/awslabs/aws-mobile-appsync-sdk-js/compare/aws-appsync-auth-link@1.0.1...aws-appsync-auth-link@2.0.1) (2019-11-15)