+++ packages/aws-appsync-subscription-link/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.0"></a>
+# [2.1.0](https://github.com/awslabs/aws-mobile-appsync-sdk-js/compare/aws-appsync-subscription-link@2.0.1...aws-appsync-subscription-link@2.1.0) (2020-04-15)
+
+
+### Features
+
+* **aws-appsync-subscription-link:** add mock support for websocket subscriptions ([#548](https://github.com/awslabs/aws-mobile-appsync-sdk-js/issues/548)) ([bac8808](https://github.com/awslabs/aws-mobile-appsync-sdk-js/commit/bac8808))
+
+
+
+
 <a name="2.0.1"></a>
 ## [2.0.1](https://github.com/awslabs/aws-mobile-appsync-sdk-js/compare/aws-appsync-subscription-link@1.0.1...aws-appsync-subscription-link@2.0.1) (2019-11-15)